Use the SHA sum with the TT-RSS images
Since I don't have a specific version number and NixOS doesn't automatically
repull the images it's better to include the SHA sum in order to update the
images while building the system. In addition it's probably better for backups
as well because the version that gets installed matches the DB backups and
doesn't pull in an incompatible one by accident.